+++ HISTORY.md
@@ -13,6 +13,9 @@
 * **Translation Updates:**
     * German
 
+### 2026-07-31
+* extensions: added "result of process" extension
+
 ### 2026-07-30
 * symbols: new animated "broken gears" symbols
 * blocks, objects: added "broken gears" symbol to visualize a process that has an error